requirements.txt
arrow==1.2.3
asgiref==3.7.2
async-timeout==4.0.3
backports.zoneinfo==0.2.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
colorama==0.4.6
cryptography==41.0.3
Django==4.2.1
django-cors-headers==4.1.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-filter==23.2
django-nose==1.4.7
django-redis==5.3.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-yasg==1.21.7
Faker==19.3.1
googlemaps==4.10.0
idna==3.4
inflection==0.5.1
Inflector==3.1.0
jsonfield==3.1.0
kafka-python==2.0.2
loguru==0.7.0
nose==1.3.7
numpy==1.24.4
packaging==23.1
pyarrow==12.0.1
pycparser==2.21
PyJWT==2.7.0
PyMySQL==1.1.0
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
querystring-parser==1.2.4
redis==4.6.0
requests==2.31.0
ruamel.yaml==0.17.32
ruamel.yaml.clib==0.2.7
six==1.16.0
sqlparse==0.4.4
typing-extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.4
win32-setctime==1.1.0
xlrd==2.0.1
xlwt==1.3.0
